Refactor VerilogRename -> RemoveKeywordCollisions
This moves VerilogRename out of Passes.scala and genericizes it as the new
Transform KemoveKeywordCollisions. This new Transform will remove keywords
for arbitrary sets of reserved keyword.

This adds VerilogRename back as a class instead of an object.
